+++ module/enrichers.js
@@ -136,10 +136,10 @@ function _onClickInlineCheck(event) {
     test.DNDescriptor = test.dn ?? (interactionAttacks.includes(test.testType) ? `target${test.testType.capitalize()}` : 'standard');
   }
 
-  if (actor.system?.skills?.[test.testType]) {
+  if (actor.system?.skills?.[test.testType] || actor.system?.customSkills?.[test.testType]) {
     // skill test
     const skillName = test.testType;
-    const skill = actor.system.skills[skillName];
+    const skill = actor.system.skills[skillName] ?? actor.system?.customSkills?.[test.testType];
     if (!skill) return ui.notifications.warn(_loc('torgeternity.notifications.noSkillNamed') + skillName);
     const attribute = actor.system.attributes[test.attribute ?? skill.baseAttribute];
     if (!attribute) return ui.notifications.warn(_loc('torgeternity.notifications.noItemNamed'));
@@ -158,6 +158,7 @@ function _onClickInlineCheck(event) {
       skillName: skillName,
       skillValue: skillValue,
       isFav: skill.isFav,
+      customSkill: !!actor.system?.customSkills?.[test.testType],
       unskilledUse: skill.unskilledUse || isInteractionAttack,
     }, { inplace: true })